Day 02 project: create moveHand() function to make code DRY.
The moveHand(hand, degrees) function handles rotating the hand to the
appropriate number of degrees clockwise from the twelve o'clock position
(straight up), while disabling the CSS transition effect if we are
crossing the threshold from just shy of 360 degrees to 0 degrees. This
prevents an unsightly effect of the hand animating backwards all the way
around from near-360 degrees back to 0 degrees.