Day 02 project: Converted CSS lengths from % to rem and other cosmetic changes.
The conversion from percentage lengths to rem lengths was for ease of
positioning.

Changed the second hand to overhand the rotation point, as second hands on real
clocks do. This necessitated tweaking the transform-origin CSS property.

Added a pin to the axis of rotation, to cover the area where the hands at times
disconnect at certain rotational positions.